requirements.txt
streamlit
openai
chardet
PyPDF2
python-docx
